# api/openapi/control-v1.yaml
openapi: 3.1.0
info:
  title: Agent OS Control API
  version: v1
  description: Public task submission and observation contract. Task status is kernel-owned.
servers:
  - url: https://control.fenced.dev
paths:
...
  /v1/workflows:
    post:
      operationId: createWorkflow
      summary: Publish a durable multi-agent workflow run (v1.2)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '202':
          description: A new workflow was accepted.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '200':
          description: An idempotency-key replay returned the existing workflow.
        '404': { description: Workflow orchestration is not configured. }
        '422': { description: The workflow document is not a valid DAG. }
  /v1/workflows/{workflowId}:
    get:
      operationId: getWorkflow
      summary: Read one workflow with its steps
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: workflowId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The workflow document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404': { description: The workflow was not found. }
  /v1/workflows/{workflowId}/cancel:
    post:
      operationId: cancelWorkflow
      summary: Record the durable cancellation intent; the orchestrator propagates it
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: workflowId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: If-Match
          in: header
          required: true
          schema: { type: string }
      responses:
        '202':
          description: The cancellation intent was recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404': { description: The workflow was not found. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
...
  schemas:
    CreateWorkflowRequest:
      type: object
      additionalProperties: false
      required: [goal, workflow]
      properties:
        namespace: { type: string, minLength: 1, maxLength: 64 }
        goal: { type: string, minLength: 1, maxLength: 8192 }
        workflow:
          type: object
          description: The workflow document (steps form a DAG; validated by the kernel).
    Workflow:
      type: object
      additionalProperties: false
      required: [id, namespace, goal, status, resourceVersion, steps, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        namespace: { type: string }
        goal: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED] }
        failureCode: { type: string }
        cancelRequestedAt: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        deadlineExceededAt: { type: string, format: date-time }
        resourceVersion: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        steps:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [name, ordinal, status, attemptCount, resourceVersion]
            properties:
              name: { type: string }
              ordinal: { type: integer }
              status: { type: string, enum: [PENDING, WAITING_APPROVAL, RUNNING, SUCCEEDED, FAILED, SKIPPED, CANCELLED] }
              attemptCount: { type: integer }
              taskId: { type: string, format: uuid }
              failureCode: { type: string }
              decidedBy: { type: string }
              approvalDecision: { type: string, enum: [approved, rejected] }
              parentStepName: { type: string }
              isDynamic: { type: boolean }
              spawnDepth: { type: integer, minimum: 1 }
              spawnKey: { type: string }
              resourceVersion: { type: integer, format: int64 }
...
security:
  - bearerAuth: []